(Fix) next_price attempt to fix dec/float issues

+++ hummingbot/strategy/pure_market_making/pure_market_making.pyx
@@ -1068,9 +1068,9 @@ cdef class PureMarketMakingStrategy(StrategyBase):
             if price_above_bid < lower_buy_price:
                 lower_buy_price = price_above_bid
             elif self._order_optimization_failsafe_enabled:
-                next_price = market.c_get_next_price(self.trading_pair, False, lower_buy_price)
-                next_price_quantum = market.c_get_order_price_quantum(self.trading_pair, next_price)
-                lower_buy_price = (ceil(next_price / next_price_quantum) + 1) * next_price_quantum
+                next_price = Decimal(market.c_get_next_price(self.trading_pair, False, lower_buy_price))
+                next_price_quantum = Decimal(market.c_get_order_price_quantum(self.trading_pair, next_price))
+                lower_buy_price = Decimal(ceil(next_price / next_price_quantum) + 1) * next_price_quantum
             ### Stonksi addition ###
 
 
@@ -1098,9 +1098,9 @@ cdef class PureMarketMakingStrategy(StrategyBase):
             if price_below_ask > higher_sell_price:
                 higher_sell_price = price_below_ask
             elif self._order_optimization_failsafe_enabled:
-                next_price = market.c_get_next_price(self.trading_pair, True, higher_sell_price)
-                next_price_quantum = market.c_get_order_price_quantum(self.trading_pair, next_price)
-                higher_sell_price = (ceil(next_price / next_price_quantum) - 1) * next_price_quantum    
+                next_price = Decimal(market.c_get_next_price(self.trading_pair, True, higher_sell_price))
+                next_price_quantum = Decimal(market.c_get_order_price_quantum(self.trading_pair, next_price))
+                higher_sell_price = Decimal(ceil(next_price / next_price_quantum) - 1) * next_price_quantum    
             ### Stonksi addition ###
 
             for i, proposed in enumerate(proposal.sells):
